roll back to the original one

+++ com/as.tool/lua/can/socketwin_can_driver.c
@@ -52,7 +52,7 @@
  */
 struct can_frame {
 	uint32_t can_id;  /* 32 bit CAN_ID + EFF/RTR/ERR flags */
-	uint32_t    can_dlc; /* frame payload length in byte (0 .. CAN_MAX_DLEN) */
+	uint8_t    can_dlc; /* frame payload length in byte (0 .. CAN_MAX_DLEN) */
 	uint8_t    data[CAN_MAX_DLEN] __attribute__((aligned(8)));
 };
 struct Can_SocketHandle_s
